http renamed to service-http
